refactor: Move auth update routes from user to auth controller
